cdk/requirements.txt
aws-cdk-lib==2.120.0
constructs>=10.0.0
aws-cdk.aws-lambda-python-alpha==2.120.0a0
